+++ src/api_info.rs
@@ -1,4 +1,4 @@
-// Copyright (C) 2019-2021 The apca Developers
+// Copyright (C) 2019-2022 The apca Developers
 // SPDX-License-Identifier: GPL-3.0-or-later
 
 use std::env::var_os;
@@ -86,5 +86,42 @@ impl ApiInfo {
       key_id,
       secret,
     })
+  }
+
+  /// Split this `ApiInfo` object back into its constituent parts.
+  ///
+  /// This method is the inverse of the [`from_parts`][Self::from_parts]
+  /// constructor. It returns a tuple comprising the base URL, key ID,
+  /// and secret.
+  pub fn into_parts(self) -> (String, String, String) {
+    let ApiInfo {
+      base_url,
+      key_id,
+      secret,
+    } = self;
+    (base_url.into(), key_id, secret)
+  }
+}
+
+
+#[cfg(test)]
+mod tests {
+  use super::*;
+
+
+  /// Check that we can create an [`ApiInfo`] object from its
+  /// constituent parts and destructure it back into them.
+  #[test]
+  fn from_into_parts() {
+    let base_url = "https://paper-api.alpaca.markets/";
+    let key_id = "XXXXXXXXXXXXXXXXXXXX";
+    let secret = "YYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYY";
+
+    let api_info = ApiInfo::from_parts(base_url, key_id, secret).unwrap();
+    let (new_base_url, new_key_id, new_secret) = api_info.into_parts();
+
+    assert_eq!(new_base_url, base_url);
+    assert_eq!(new_key_id, key_id);
+    assert_eq!(new_secret, secret);
   }
 }